requirements.txt
PyMySQL==1.0.2
SQLAlchemy==1.4.43
uvicorn==0.18.3
pydantic==1.10.2
virtualenv==20.16.6
dotenv-cli==3.1.0
python-dotenv==0.21.0
fastapi==0.85.1